Move code related to devServer into dev_server.js
The logic for mapping keys in the shakapacker.yml file is straightforward: if
a key is specific to webpack, it will be mapped to camelCase, otherwise, the
key will remain as is. This means that any other key that is not specific to
webpack (such as random_key) will remain accessible in its original form.